refactor: create order route
